Make SDC library widget tests use Robolectric (instead of running on emulator)
