+++ datacapture/src/test/java/com/google/android/fhir/datacapture/views/QuestionnaireItemCheckBoxGroupViewHolderFactoryTest.kt
@@ -17,42 +17,43 @@
 package com.google.android.fhir.datacapture.views
 
 import android.view.ViewGroup
+import android.widget.CheckBox
 import android.widget.FrameLayout
-import android.widget.RadioButton
 import android.widget.TextView
-import androidx.appcompat.view.ContextThemeWrapper
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.view.children
-import androidx.test.annotation.UiThreadTest
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
 import com.google.android.fhir.datacapture.ChoiceOrientationTypes
 import com.google.android.fhir.datacapture.EXTENSION_CHOICE_ORIENTATION_URL
+import com.google.android.fhir.datacapture.EXTENSION_OPTION_EXCLUSIVE_URL
 import com.google.android.fhir.datacapture.R
 import com.google.common.truth.Truth.assertThat
+import org.hl7.fhir.r4.model.BooleanType
 import org.hl7.fhir.r4.model.CodeType
 import org.hl7.fhir.r4.model.Coding
+import org.hl7.fhir.r4.model.Extension
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.RuntimeEnvironment
 
-@RunWith(AndroidJUnit4::class)
-class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
+@RunWith(RobolectricTestRunner::class)
+class QuestionnaireItemCheckBoxGroupViewHolderFactoryTest {
   private val parent =
     FrameLayout(
-      ContextThemeWrapper(
-        InstrumentationRegistry.getInstrumentation().targetContext,
-        R.style.Theme_MaterialComponents
-      )
+      RuntimeEnvironment.getApplication().apply { setTheme(R.style.Theme_MaterialComponents) }
     )
-  private val viewHolder = QuestionnaireItemRadioGroupViewHolderFactory.create(parent)
+  private val viewHolder = QuestionnaireItemCheckBoxGroupViewHolderFactory.create(parent)
 
   @Test
   fun bind_shouldSetQuestionHeader() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { text = "Question?" },
+        Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
+          text = "Question?"
+        },
         QuestionnaireResponse.QuestionnaireResponseItemComponent()
       ) {}
     )
@@ -62,9 +63,10 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun bind_vertical_shouldCreateRadioButtons() {
+  fun bind_vertical_shouldCreateCheckBoxButtons() {
     val questionnaire =
       Questionnaire.QuestionnaireItemComponent().apply {
+        repeats = true
         addExtension(
           EXTENSION_CHOICE_ORIENTATION_URL,
           CodeType(ChoiceOrientationTypes.VERTICAL.extensionCode)
@@ -87,8 +89,8 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
       ) {}
     )
 
-    val radioGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group)
-    val children = radioGroup.children.asIterable().filterIsInstance<RadioButton>()
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    val children = checkBoxGroup.children.asIterable().filterIsInstance<CheckBox>()
     children.forEachIndexed { index, view ->
       assertThat(view.text).isEqualTo(questionnaire.answerOption[index].valueCoding.display)
       assertThat(view.layoutParams.width).isEqualTo(ViewGroup.LayoutParams.MATCH_PARENT)
@@ -96,9 +98,10 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun bind_horizontal_shouldCreateRadioButtons() {
+  fun bind_horizontal_shouldCreateCheckBoxButtons() {
     val questionnaire =
       Questionnaire.QuestionnaireItemComponent().apply {
+        repeats = true
         addExtension(
           EXTENSION_CHOICE_ORIENTATION_URL,
           CodeType(ChoiceOrientationTypes.HORIZONTAL.extensionCode)
@@ -121,8 +124,8 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
       ) {}
     )
 
-    val radioGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group)
-    val children = radioGroup.children.asIterable().filterIsInstance<RadioButton>()
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    val children = checkBoxGroup.children.asIterable().filterIsInstance<CheckBox>()
     children.forEachIndexed { index, view ->
       assertThat(view.text).isEqualTo(questionnaire.answerOption[index].valueCoding.display)
       assertThat(view.layoutParams.width).isEqualTo(ViewGroup.LayoutParams.WRAP_CONTENT)
@@ -130,10 +133,11 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun bind_noAnswer_shouldLeaveRadioButtonsUnchecked() {
+  fun bind_noAnswer_shouldLeaveCheckButtonsUnchecked() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
           addAnswerOption(
             Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
               value = Coding().apply { display = "Coding 1" }
@@ -144,96 +148,177 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
       ) {}
     )
 
-    val radioButton =
-      viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(1) as
-        RadioButton
-    assertThat(radioButton.isChecked).isFalse()
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    val checkBox = checkBoxGroup.getChildAt(1) as CheckBox
+    assertThat(checkBox.isChecked).isFalse()
   }
 
   @Test
-  @UiThreadTest
-  fun bind_answer_shouldCheckRadioButton() {
+  fun bind_answer_shouldSetCheckBoxButton() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
           addAnswerOption(
             Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-              value = Coding().apply { display = "Coding 1" }
-            }
-          )
-          addAnswerOption(
-            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-              value = Coding().apply { display = "Coding 2" }
+              value =
+                Coding().apply {
+                  code = "code 1"
+                  display = "Coding 1"
+                }
             }
           )
         },
         QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
           addAnswer(
             QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = Coding().apply { display = "Coding 1" }
+              value =
+                Coding().apply {
+                  code = "code 1"
+                  display = "Coding 1"
+                }
             }
           )
         }
       ) {}
     )
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    val checkBox = checkBoxGroup.getChildAt(1) as CheckBox
 
-    assertThat(
-        (viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(1) as
-            RadioButton)
-          .isChecked
-      )
-      .isTrue()
-    assertThat(
-        (viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(2) as
-            RadioButton)
-          .isChecked
-      )
-      .isFalse()
+    assertThat(checkBox.isChecked).isTrue()
   }
 
   @Test
-  @UiThreadTest
-  fun click_shouldSetQuestionnaireResponseItemAnswer() {
+  fun click_shouldAddQuestionnaireResponseItemAnswer() {
     val questionnaireItemViewItem =
       QuestionnaireItemViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
           addAnswerOption(
             Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-              value = Coding().apply { display = "Coding 1" }
+              value =
+                Coding().apply {
+                  code = "code 1"
+                  display = "Coding 1"
+                }
             }
           )
         },
         QuestionnaireResponse.QuestionnaireResponseItemComponent()
       ) {}
     viewHolder.bind(questionnaireItemViewItem)
-    viewHolder
-      .itemView
-      .findViewById<ConstraintLayout>(R.id.radio_group)
-      .getChildAt(1)
-      .performClick()
-
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    val checkBox = checkBoxGroup.getChildAt(1) as CheckBox
+    checkBox.performClick()
     val answer = questionnaireItemViewItem.questionnaireResponseItem.answer
-    assertThat(answer.size).isEqualTo(1)
+
+    assertThat(answer).hasSize(1)
     assertThat(answer[0].valueCoding.display).isEqualTo("Coding 1")
   }
 
   @Test
-  @UiThreadTest
-  fun click_shouldCheckRadioButton() {
+  fun optionExclusiveAnswerOption_click_deselectsOtherAnswerOptions() {
     val questionnaireItemViewItem =
       QuestionnaireItemViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
+          addAnswerOption(
+            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
+              value =
+                Coding().apply {
+                  code = "code-1"
+                  display = "display-1"
+                }
+              extension = listOf(Extension(EXTENSION_OPTION_EXCLUSIVE_URL, BooleanType(true)))
+            }
+          )
+          addAnswerOption(
+            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
+              value =
+                Coding().apply {
+                  code = "code-2"
+                  display = "display-2"
+                }
+            }
+          )
+        },
+        QuestionnaireResponse.QuestionnaireResponseItemComponent()
+      ) {}
+    viewHolder.bind(questionnaireItemViewItem)
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    (checkBoxGroup.getChildAt(2) as CheckBox).performClick()
+    (checkBoxGroup.getChildAt(1) as CheckBox).performClick()
+    val answer = questionnaireItemViewItem.questionnaireResponseItem.answer
+
+    assertThat(answer.single().valueCoding.display).isEqualTo("display-1")
+  }
+
+  @Test
+  fun answerOption_click_deselectsOptionExclusiveAnswerOption() {
+    val questionnaireItemViewItem =
+      QuestionnaireItemViewItem(
+        Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
+          addAnswerOption(
+            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
+              value =
+                Coding().apply {
+                  code = "code-1"
+                  display = "display-1"
+                }
+              extension = listOf(Extension(EXTENSION_OPTION_EXCLUSIVE_URL, BooleanType(true)))
+            }
+          )
+          addAnswerOption(
+            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
+              value =
+                Coding().apply {
+                  code = "code-2"
+                  display = "display-2"
+                }
+            }
+          )
+        },
+        QuestionnaireResponse.QuestionnaireResponseItemComponent()
+      ) {}
+
+    viewHolder.bind(questionnaireItemViewItem)
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    (checkBoxGroup.getChildAt(1) as CheckBox).performClick()
+    (checkBoxGroup.getChildAt(2) as CheckBox).performClick()
+    val answer = questionnaireItemViewItem.questionnaireResponseItem.answer
+
+    assertThat(answer.single().valueCoding.display).isEqualTo("display-2")
+  }
+
+  @Test
+  fun click_shouldRemoveQuestionnaireResponseItemAnswer() {
+    val questionnaireItemViewItem =
+      QuestionnaireItemViewItem(
+        Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
           answerValueSet = "http://coding-value-set-url"
         },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
+        QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
+          addAnswer(
+            QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
+              value =
+                Coding().apply {
+                  code = "code 1"
+                  display = "Coding 1"
+                }
+            }
+          )
+        },
         resolveAnswerValueSet = {
           if (it == "http://coding-value-set-url") {
             listOf(
               Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-                value = Coding().apply { display = "Coding 1" }
-              },
-              Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-                value = Coding().apply { display = "Coding 2" }
+                value =
+                  Coding().apply {
+                    code = "code 1"
+                    display = "Coding 1"
+                  }
               }
             )
           } else {
@@ -242,80 +327,22 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
         }
       ) {}
     viewHolder.bind(questionnaireItemViewItem)
-    viewHolder
-      .itemView
-      .findViewById<ConstraintLayout>(R.id.radio_group)
-      .getChildAt(1)
-      .performClick()
+    val checkBoxGroup = viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group)
+    val checkBox = checkBoxGroup.getChildAt(1) as CheckBox
+    checkBox.performClick()
+    val answer = questionnaireItemViewItem.questionnaireResponseItem.answer
 
-    assertThat(
-        (viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(1) as
-            RadioButton)
-          .isChecked
-      )
-      .isTrue()
-    assertThat(
-        (viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(2) as
-            RadioButton)
-          .isChecked
-      )
-      .isFalse()
+    assertThat(answer).isEmpty()
   }
 
   @Test
-  @UiThreadTest
-  fun click_shouldCheckOtherRadioButton() {
-    viewHolder.bind(
-      QuestionnaireItemViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          addAnswerOption(
-            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-              value = Coding().apply { display = "Coding 1" }
-            }
-          )
-          addAnswerOption(
-            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-              value = Coding().apply { display = "Coding 2" }
-            }
-          )
-        },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
-          addAnswer(
-            QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = Coding().apply { display = "Coding 1" }
-            }
-          )
-        }
-      ) {}
-    )
-
-    assertThat(
-        (viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(1) as
-            RadioButton)
-          .isChecked
-      )
-      .isTrue()
-
-    viewHolder
-      .itemView
-      .findViewById<ConstraintLayout>(R.id.radio_group)
-      .getChildAt(2)
-      .performClick()
-
-    assertThat(
-        (viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(2) as
-            RadioButton)
-          .isChecked
-      )
-      .isTrue()
-  }
-
-  @Test
-  @UiThreadTest
   fun displayValidationResult_error_shouldShowErrorMessage() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { required = true },
+        Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
+          required = true
+        },
         QuestionnaireResponse.QuestionnaireResponseItemComponent()
       ) {}
     )
@@ -325,11 +352,11 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun displayValidationResult_noError_shouldShowNoErrorMessage() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
           required = true
           addAnswerOption(
             Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
@@ -355,6 +382,7 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
     viewHolder.bind(
       QuestionnaireItemViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
+          repeats = true
           readOnly = true
           addAnswerOption(
             Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
@@ -365,10 +393,12 @@ class QuestionnaireItemRadioGroupViewHolderFactoryInstrumentedTest {
         QuestionnaireResponse.QuestionnaireResponseItemComponent()
       ) {}
     )
-    val radioButton =
-      viewHolder.itemView.findViewById<ConstraintLayout>(R.id.radio_group).getChildAt(1) as
-        RadioButton
 
-    assertThat(radioButton.isEnabled).isFalse()
+    assertThat(
+        (viewHolder.itemView.findViewById<ConstraintLayout>(R.id.checkbox_group).getChildAt(1) as
+            CheckBox)
+          .isEnabled
+      )
+      .isFalse()
   }
 }
