+++ datacapture/src/test/java/com/google/android/fhir/datacapture/views/QuestionnaireItemDropDownViewHolderFactoryTest.kt
@@ -19,10 +19,6 @@ package com.google.android.fhir.datacapture.views
 import android.widget.AutoCompleteTextView
 import android.widget.FrameLayout
 import android.widget.TextView
-import androidx.appcompat.view.ContextThemeWrapper
-import androidx.test.annotation.UiThreadTest
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
 import com.google.android.fhir.datacapture.R
 import com.google.android.fhir.datacapture.displayString
 import com.google.android.material.textfield.TextInputLayout
@@ -32,30 +28,20 @@ import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.hl7.fhir.r4.model.StringType
-import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.RuntimeEnvironment
 
-@RunWith(AndroidJUnit4::class)
-class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
-  private lateinit var context: ContextThemeWrapper
-  private lateinit var parent: FrameLayout
-  private lateinit var viewHolder: QuestionnaireItemViewHolder
-
-  @Before
-  @UiThreadTest
-  fun setUp() {
-    context =
-      ContextThemeWrapper(
-        InstrumentationRegistry.getInstrumentation().targetContext,
-        R.style.Theme_MaterialComponents
-      )
-    parent = FrameLayout(context)
-    viewHolder = QuestionnaireItemDropDownViewHolderFactory.create(parent)
-  }
+@RunWith(RobolectricTestRunner::class)
+class QuestionnaireItemDropDownViewHolderFactoryTest {
+  private val parent =
+    FrameLayout(
+      RuntimeEnvironment.getApplication().apply { setTheme(R.style.Theme_MaterialComponents) }
+    )
+  private val viewHolder = QuestionnaireItemDropDownViewHolderFactory.create(parent)
 
   @Test
-  @UiThreadTest
   fun shouldSetQuestionHeader() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
@@ -69,7 +55,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun shouldPopulateDropDown() {
     val answerOption =
       Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
@@ -94,7 +79,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun shouldSetDropDownOptionToCodeIfValueCodingDisplayEmpty() {
     val answerOption =
       Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
@@ -119,7 +103,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun shouldSetAutoTextViewEmptyIfAnswerNull() {
     val answerOption =
       Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
@@ -143,7 +126,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun shouldAutoCompleteTextViewToDisplayIfAnswerNotNull() {
     val answerOption =
       Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
@@ -182,7 +164,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun shouldThrowErrorForAnswerOptionWithoutCoding() {
     val answerOption =
       Questionnaire.QuestionnaireItemAnswerOptionComponent().apply { value = StringType("test") }
@@ -204,7 +185,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun displayValidationResult_error_shouldShowErrorMessage() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
@@ -218,7 +198,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun displayValidationResult_noError_shouldShowNoErrorMessage() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
@@ -244,7 +223,6 @@ class QuestionnaireItemDropDownViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  @UiThreadTest
   fun bind_readOnly_shouldDisableView() {
     viewHolder.bind(
       QuestionnaireItemViewItem(
